Initial go at moving get_file_size() to caiman.base.movies
